requirements.txt
# Core dependencies
gitpython>=3.1.40
pyyaml>=6.0.1
aiofiles>=23.2.1

# Database
aiosqlite>=0.19.0

# Validation
jsonschema>=4.20.0

# CLI & UI
rich>=13.7.0
click>=8.1.7

# Utilities
python-dotenv>=1.0.0
